feat: modify migration to define potencia and economia as DECIMAL(8,2)

+++ src/main/resources/db/migration/V1__CREATE-TABLE-VEHICLE.sql
@@ -3,8 +3,8 @@ CREATE TABLE Vehicles (
      marca VARCHAR(255) NOT NULL,
      modelo VARCHAR(255) NOT NULL,
      ano INT NOT NULL,
-     potencia DOUBLE PRECISION NOT NULL,
-     economia DOUBLE PRECISION NOT NULL,
+     potencia DECIMAL(8, 2) NOT NULL,
+     economia DECIMAL(8, 2) NOT NULL,
      tipo VARCHAR(50) NOT NULL,
      preco DECIMAL(15,2) NOT NULL
 );
